reorganize provider logic [skip ci]
