+++ src/lib/ai/providers/custom-provider-factory.ts
