+++ src/lib/ai/providers/factories/cerebras.ts
