+++ src/lib/ai/providers/factories/groq.ts
