+++ src/lib/ai/providers/factories/openrouter.ts
